# docs/openapi-consumer.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "X-Ray 企业变化解释器",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v2/discovery": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Discovery",
        "description": "Network discovery is opt-in; no request writes verified company evidence.",
        "operationId": "discovery_api_v2_discovery_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 80,
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consumer/discovery": {
      "post": {
        "tags": [
          "消费者实时调查"
        ],
        "summary": "Discover",
        "operationId": "discover_api_consumer_discovery_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas__consumer__DiscoveryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/company-research/discovery": {
      "post": {
        "tags": [
          "company research"
        ],
        "summary": "Discovery",
        "operationId": "discovery_api_company_research_discovery_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas__company_research__DiscoveryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/regions": {
      "get": {
        "summary": "Regions",
        "operationId": "regions_api_regions_get",
        "parameters": [
          {
            "name": "parent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{6}$",
              "default": "100000",
              "title": "Parent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/places": {
      "get": {
        "summary": "Places",
        "operationId": "places_api_places_get",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 100,
              "title": "Keyword"
            }
          },
          {
            "name": "region_code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(\\d{6})?$",
              "default": "",
              "title": "Region Code"
            }
          },
          {
            "name": "street",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 50,
              "default": "",
              "title": "Street"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/legal-entities": {
      "get": {
        "summary": "Legal Entities",
        "operationId": "legal_entities_api_legal_entities_get",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 100,
              "title": "Keyword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/company-report": {
      "post": {
        "summary": "Company Report",
        "operationId": "company_report_api_company_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyReportInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyReportInput": {
        "properties": {
          "company_keyword": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Company Keyword"
          },
          "place": {
            "$ref": "#/components/schemas/SelectedPlace"
          }
        },
        "type": "object",
        "required": [
          "company_keyword",
          "place"
        ],
        "title": "CompanyReportInput"
      },
...
      "Discovery": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "title": "Investigation Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "mode": {
            "type": "string",
            "enum": [
              "live_search",
              "cached",
              "unavailable"
            ],
            "title": "Mode"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/IdentityCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Sources"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "type": "array",
            "title": "Trace"
          },
          "unknowns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknowns"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "investigation_id",
          "query",
          "location",
          "generated_at",
          "mode",
          "candidates",
          "sources",
          "trace",
          "unknowns"
        ],
        "title": "Discovery"
      },
      "DiscoveryResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "companies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Companies"
          },
          "announcements": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Announcements"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "external_search_links": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "External Search Links"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "status",
          "query",
          "companies",
          "announcements",
          "verification_status",
          "external_search_links",
          "limitations"
        ],
        "title": "DiscoveryResponse"
      },
...
      "Evidence": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "publisher": {
            "type": "string",
            "title": "Publisher"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "date_semantics": {
            "type": "string",
            "title": "Date Semantics",
            "default": "公开日期未核实"
          },
          "fetched_at": {
            "type": "string",
            "title": "Fetched At"
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "search_excerpt",
              "page_text",
              "provider_response"
            ],
            "title": "Verification Status",
            "default": "search_excerpt"
          },
          "channel": {
            "type": "string",
            "enum": [
              "public_web",
              "community",
              "government",
              "registry"
            ],
            "title": "Channel",
            "default": "public_web"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "page_status": {
            "type": "string",
            "title": "Page Status",
            "default": "未读取原网页"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          },
          "scope": {
            "type": "string",
            "enum": [
              "selected_entity",
              "brand_context"
            ],
            "title": "Scope",
            "default": "selected_entity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "publisher",
          "excerpt",
          "fetched_at",
          "purpose"
        ],
        "title": "Evidence"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdentityCandidate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "basis": {
            "type": "string",
            "title": "Basis"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "relationship_status": {
            "type": "string",
            "title": "Relationship Status",
            "default": "网页提及，门店关联尚未核实"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "basis",
          "source_ids"
        ],
        "title": "IdentityCandidate"
      },
...
      "SelectedPlace": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "address": {
            "type": "string",
            "maxLength": 500,
            "title": "Address",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "SelectedPlace"
      },
...
      "Step": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action",
          "status",
          "detail"
        ],
        "title": "Step"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__schemas__company_research__DiscoveryInput": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "pattern": "^[^\\x00-\\x1f\\x7f]+$",
            "title": "Query"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "DiscoveryInput"
      },
      "app__schemas__consumer__DiscoveryInput": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "pattern": "^[^\\x00-\\x1f\\x7f]+$",
            "title": "Query"
          },
          "location": {
            "type": "string",
            "maxLength": 60,
            "pattern": "^[^\\x00-\\x1f\\x7f]*$",
            "title": "Location",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "DiscoveryInput"
      }
    }
  }
}